+++ src/git/GitRepository.ts
@@ -1,23 +1,23 @@
 import createRepository, { SimpleGit as Repository, StatusResult, DefaultLogFields } from 'simple-git';
 import { join } from 'path';
 import { pathExists, rmdir, emptyDir, ensureDir } from 'fs-extra';
-import conventionalCommitsParser from 'conventional-commits-parser';
 import { GitLog } from './GitLog';
-import { Readable } from 'stream';
+import { ProjectConfig } from '../configs';
+import { Utils } from '../tools';
 
 /**
  * A simple API with basic functionality of a git repository.
  */
 export class GitRepository {
-  protected repoPath: string;
+  protected config?: ProjectConfig;
 
   /**
    * Initializes a new instance of this class.
    *
-   * @param repoPath - Path of the git repository.
+   * @param config - The project configuration.
    */
-  constructor(repoPath?: string) {
-    this.repoPath = repoPath ?? process.cwd();
+  constructor(config?: ProjectConfig) {
+    this.config = config;
   }
 
   /**
@@ -26,15 +26,15 @@ export class GitRepository {
    * @returns The path to the git repository.
    */
   public getRepoPath(): string {
-    return this.repoPath;
+    return this.config?.projectPath ?? process.cwd();
   }
 
   /**
    * Gets the path of the git repository.
    */
   public async remove(): Promise<void> {
-    await emptyDir(this.repoPath);
-    await rmdir(this.repoPath);
+    await emptyDir(this.getRepoPath());
+    await rmdir(this.getRepoPath());
   }
 
   /**
@@ -115,27 +115,11 @@ export class GitRepository {
    * Collects all commit messages since the last release.
    */
   public async getLogsSinceLastRelease(): Promise<GitLog[]> {
-    const gitLogs: GitLog[] = [];
     const logs = await this.getDiffLogs();
     const logMessages = logs.map((log) => `${log.message}\n\n${log.body}\n\n${log.refs}`);
-    const stream = Readable.from(logMessages);
-    let currIndex = 0;
-    return new Promise((resolve, reject) => {
-      stream
-        .pipe(conventionalCommitsParser())
-        .on('error', function (err) {
-          err.message = 'Error in conventional-commits-parser: ' + err.message;
-          reject(err);
-        })
-        .on('data', (data) => {
-          const log = data as GitLog;
-          log.hash = logs[currIndex++].hash;
-          gitLogs.push(log);
-        })
-        .on('end', () => {
-          resolve(gitLogs);
-        });
-    });
+    const gitLogs = await Utils.parseConventionalCommits(logMessages, this.config?.conventionalCommit);
+    gitLogs.forEach((v, i) => (v.hash = logs[i].hash));
+    return gitLogs;
   }
 
   private async getDiffLogs(): Promise<readonly DefaultLogFields[]> {
@@ -151,12 +135,13 @@ export class GitRepository {
 
   private async createOrOpenRepo(): Promise<Repository> {
     let repo: Repository;
-    const gitFolder = join(this.repoPath, '.git');
-    await ensureDir(this.repoPath);
+    const path = this.getRepoPath();
+    const gitFolder = join(path, '.git');
+    await ensureDir(path);
     if (await pathExists(gitFolder)) {
-      repo = createRepository(this.repoPath);
+      repo = createRepository(path);
     } else {
-      repo = createRepository(this.repoPath);
+      repo = createRepository(path);
       await repo.init();
     }
     return repo;