+++ doc/modules/changelog.md
@@ -13,8 +13,10 @@
 * [ChangelogWriter](../classes/changelog.changelogwriter.md)
 * [ChangelogWriterFactory](../classes/changelog.changelogwriterfactory.md)
 * [ConventionalChangelogWriter](../classes/changelog.conventionalchangelogwriter.md)
+* [KeepAChangelogWriter](../classes/changelog.keepachangelogwriter.md)
 
 ### Interfaces
 
 * [ChangelogWriterOptions](../interfaces/changelog.changelogwriteroptions.md)
 * [ConventionalChangelogWriterOptions](../interfaces/changelog.conventionalchangelogwriteroptions.md)
+* [KeepAChangelogWriterOptions](../interfaces/changelog.keepachangelogwriteroptions.md)